feat: admin page

See merge request xhoschek/pb138-student-organiser!56

+++ backend/apis/user/controller.ts
@@ -9,14 +9,15 @@ import {
 } from "./validationSchemas";
 
 import {notebookRepository} from "../notebook/repository";
+import {flashdeckCreateRequestSchema, getAllDecksSchema} from "../deck/validationSchemas";
+import {deckRepository} from "../deck/repository";
 
 
 const get = async (req: Request, res: Response): Promise<void> => {
-    if (!req.params.userId) {
-        res.status(400).send("No user Id provided");
-        return;
-    }
-    let maybeUser = await userRepository.findById(req.params.userId);
+    const userId = req.session.passport?.user.id;
+    if (!userId) { return; }
+
+    const maybeUser = await userRepository.findById(userId);
     if (maybeUser.isErr) {
         handleRepositoryErrors(maybeUser.error, res);
         return;
@@ -26,12 +27,12 @@ const get = async (req: Request, res: Response): Promise<void> => {
 }
 
 const post = async (req: Request, res: Response) => {
-    let request = await parseRequest(userCreateRequestSchema, req, res);
+    const request = await parseRequest(userCreateRequestSchema, req, res);
     if (!request) {
         return;
     }
 
-    let maybeUser = await userRepository.create(request);
+    const maybeUser = await userRepository.create(request);
     if (maybeUser.isErr) {
         handleRepositoryErrors(maybeUser.error, res)
         return;
@@ -40,23 +41,37 @@ const post = async (req: Request, res: Response) => {
     res.status(200).send(maybeUser.unwrap());
 }
 const put = async (req: Request, res: Response) => {
-    let request = await parseRequest(userUpdateRequestSchema, req, res);
-    if (!request) {
+    const userId = req.session.passport?.user.id;
+    const request = await parseRequest(userUpdateRequestSchema, req, res);
+    if (!request || !userId) {
         return;
     }
-    let maybeUser = await userRepository.update(request);
+    const maybeUser = await userRepository.update(request, userId);
     if (maybeUser.isErr) {
         handleRepositoryErrors(maybeUser.error, res);
         return;
     }
     res.status(200).send(maybeUser.unwrap());
 }
+
 const remove = async (req: Request, res: Response) => {
-    if (!req.params.userId) {
+    const userId = req.session.passport?.user.id;
+    if (!userId) {
         res.status(400).send("No user Id provided");
         return;
     }
-    let maybeUser = await userRepository.delete(req.params.userId);
+    const admin = await userRepository.isAdmin(userId);
+    if (admin.isErr) {
+        handleRepositoryErrors(admin.error, res);
+        return;
+    }
+
+    if (!admin.unwrap().isAdmin || req.session.passport?.user.id == req.params.userId){
+        res.status(401).send("Unauthorized");
+        return;
+    }
+
+    const maybeUser = await userRepository.delete(req.params.userId);
     if (maybeUser.isErr) {
         handleRepositoryErrors(maybeUser.error, res);
         return;
@@ -65,15 +80,13 @@ const remove = async (req: Request, res: Response) => {
 }
 
 const getNotebooks = async (req: Request, res: Response) => {
-    let request = await parseRequest(userGetNotebooksRequestSchema, req, res);
-    if (!request) {
+    const userId = req.session.passport?.user.id;
+    const request = await parseRequest(userGetNotebooksRequestSchema, req, res);
+    if (!request || !userId) {
         return;
     }
 
-    let notebooks = await notebookRepository.getAll({
-        userId: request.params.userId,
-        withTags: request.query.withTags
-    });
+    const notebooks = await notebookRepository.getAll(request.query.withTags, userId);
     if (notebooks.isErr) {
         handleRepositoryErrors(notebooks.error, res);
         return;
@@ -82,17 +95,46 @@ const getNotebooks = async (req: Request, res: Response) => {
 }
 
 const createNotebook = async (req: Request, res: Response) => {
-    let request = await parseRequest(userCreateNotebookRequestSchema, req, res);
-    if (!request) {
+    const userId = req.session.passport?.user.id;
+    const request = await parseRequest(userCreateNotebookRequestSchema, req, res);
+    if (!request || !userId) {
         return;
     }
-    let notebook = await notebookRepository.createFromUser(request);
+    const notebook = await notebookRepository.createFromUser(request, userId);
     if (notebook.isErr) {
         handleRepositoryErrors(notebook.error, res);
         return;
     }
     res.status(200).send(notebook.unwrap());
+}
 
+const createDeck = async (req: Request, res: Response) => {
+    const userId = req.session.passport?.user.id;
+    const request = await parseRequest(flashdeckCreateRequestSchema, req, res);
+    if (!request || !userId) {
+        return;
+    }
+
+    const flashdeck = await deckRepository.create(request, userId);
+    if (flashdeck.isErr) {
+        handleRepositoryErrors(flashdeck.error, res);
+        return;
+    }
+    res.status(200).send(flashdeck.unwrap());
+}
+
+const getDecks = async (req: Request, res: Response) => {
+    const userId = req.session.passport?.user.id;
+    const request = await parseRequest(getAllDecksSchema, req, res);
+    if (!userId || !request) {
+        return;
+    }
+    const decks = await deckRepository.getAll(request.query.withTags, userId);
+    if (decks.isErr) {
+        handleRepositoryErrors(decks.error, res);
+        return;
+    }
+    res.status(200).send(decks.unwrap());
 }
 
 export const userController = {
@@ -102,4 +144,6 @@ export const userController = {
     remove,
     getNotebooks,
     createNotebook,
+    getDecks,
+    createDeck
 }
